修改通关逻辑

+++ Assets/Scripts/Luban/PlayerConfignew_PlayerConfig.cs
@@ -47,7 +47,7 @@ public sealed partial class PlayerConfignew_PlayerConfig : Luban.BeanBase
     /// <summary>
     /// 当前等级最高子弹价格（等同伤害））
     /// </summary>
-    public readonly long Total;
+    public readonly float Total;
     /// <summary>
     /// 当前等级玩家血量（用不到））
     /// </summary>
